requirements.txt
Pillow==9.5.0
torch
torchvision
scikit-learn
matplotlib